Extract `Plugin` into `plugin.ts`
